requirements.txt
wsgidav>=4.3.3
python-dotenv>=1.0.0
requests>=2.32.3
waitress>=3.0.2
cachetools>=5.3.3
lxml>=5.3.0
